Move regression_test from /srvs to /cmd
